step 10
Persistence done using a file that is stored in the server.
It seems that with Sinatra sessions can't be used to transfer
the status of Temperature and PowerSavingMode into different views,
in the Post where the data lands, we need to send that data directly
onto the database, or as we have done use a file to get the data
persisting.